+++ packages/block-library/src/image/index.php
@@ -47,6 +47,18 @@ function render_block_core_image( $attributes, $content, $block ) {
 		isset( $lightbox_settings['enabled'] ) &&
 		true === $lightbox_settings['enabled']
 	) {
+		$suffix = wp_scripts_get_suffix();
+		if ( defined( 'IS_GUTENBERG_PLUGIN' ) && IS_GUTENBERG_PLUGIN ) {
+			$module_url = gutenberg_url( "/build/interactivity/image{$suffix}.js" );
+		}
+
+		wp_register_script_module(
+			'@wordpress/block-library/image',
+			isset( $module_url ) ? $module_url : includes_url( "blocks/image/view{$suffix}.js" ),
+			array( '@wordpress/interactivity' ),
+			defined( 'GUTENBERG_VERSION' ) ? GUTENBERG_VERSION : get_bloginfo( 'version' )
+		);
+
 		wp_enqueue_script_module( '@wordpress/block-library/image' );
 
 		/*
@@ -321,13 +333,6 @@ function register_block_core_image() {
 		array(
 			'render_callback' => 'render_block_core_image',
 		)
-	);
-
-	wp_register_script_module(
-		'@wordpress/block-library/image',
-		defined( 'IS_GUTENBERG_PLUGIN' ) && IS_GUTENBERG_PLUGIN ? gutenberg_url( '/build/interactivity/image.min.js' ) : includes_url( 'blocks/image/view.min.js' ),
-		array( '@wordpress/interactivity' ),
-		defined( 'GUTENBERG_VERSION' ) ? GUTENBERG_VERSION : get_bloginfo( 'version' )
 	);
 }
 add_action( 'init', 'register_block_core_image' );